Latest version
git-svn-id: http://hdt-it.googlecode.com/svn/trunk@173 ca5865f0-ea67-0c57-7e18-94a3701c88e0

+++ hdt-it/Camera.cpp
@@ -105,8 +105,8 @@ QPointF Camera::getRotation() {
 
 void Camera::rotateCamera(double x,double y){
     rotationAnimation.stop();
-    rotx += (float) 0.5f * x;
-    roty += (float) 0.5f * y;
+    rotx += 0.5 * x;
+    roty += 0.5 * y;
     emit rotationChanged();
 }
